// silap-colmirwcl/src/app/modules/silapapi/model/parametro.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Parametro { 
    codParametro?: string;
    dFineValidita?: Date;
    dInizioValidita?: Date;
    dsNote?: string;
    dsValore?: string;
    flgTipoValore?: string;
    idEsoTParametro?: number;
}