// silap-colmirwcl/src/app/modules/silapapi/model/versamento.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Ccnl } from './ccnl';
import { CreditoResiduo } from './creditoResiduo';
import { PosizioneDebitoria } from './posizioneDebitoria';
import { VersamentoProspetto } from './versamentoProspetto';
import { VersamentoProvincia } from './versamentoProvincia';
import { VersamentoSede } from './versamentoSede';
import { VersamentoStato } from './versamentoStato';

export interface Versamento { 
    annoProtocollo?: number;
    annoRiferimento?: number;
    codFiscale?: string;
    codFiscaleSoggetto?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    dProtocollo?: Date;
    dsDenominazioneAzienda?: string;
    dsEmailRiferimento?: string;
    esoTCreditoResiduo?: CreditoResiduo;
    esoTPosizioneDebitorias?: Array<PosizioneDebitoria>;
    esoTVersamentoProspettos?: Array<VersamentoProspetto>;
    esoTVersamentoProvincias?: Array<VersamentoProvincia>;
    esoTVersamentoStatos?: Array<VersamentoStato>;
    idEsoTVersamento?: number;
    idSilAziAnagrafica?: number;
    nTimestamp?: number;
    numCreditoResiduo?: number;
    numProtocollo?: number;
    numRate?: number;
    partivaIva?: string;
    printCalcoloAutomatico?: string;
    silapDCcnl?: Ccnl;
    versamentoSede?: VersamentoSede;
}